+++ CHANGELOG.md
@@ -41,6 +41,11 @@
 - "textDocument/didSave" is implemented
 - "classInfo" is now "serenata/deprecated/getClassInfo"
 - "classList" is now "serenata/deprecated/getClassListForFile"
+- "deduceTypes" is now "serenata/deprecated/deduceTypes"
+- "globalConstants" is now "serenata/deprecated/getGlobalConstants"
+- "globalFunctions" is now "serenata/deprecated/getGlobalFunctions"
+- "resolveType" is now "serenata/deprecated/resolveType"
+- "localizeType" is now "serenata/deprecated/localizeType"
 - "reindexProgressInformation" was renamed to "serenata/didProgressIndexing" and is now follows the notification instead of the response format
     - TODO: Should document that this custom notification exists on the wiki
 - In output data of extension commands (non-LSP commands), "filename" was renamed to "uri" and is now an actual URI instead of just a path